streambeats/requirements/base.txt
Django==1.6.5
South==1.0
markdown==2.4.1
django-filter==0.7
djangorestframework==2.3.14
pillow==2.5.0
